LayoutTests/media/modern-media-controls/macos-inline-media-controls/macos-inline-media-controls-controls-bar-styles-expected.txt

Testing the MacOSInlineMediaControls computed styles.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS style.left is "0px"
PASS style.bottom is "0px"
PASS style.width is "680px"
PASS style.height is "50px"
PASS bounds.left is 0
PASS bounds.top is 250
PASS bounds.width is 680
PASS bounds.height is 50

PASS successfullyParsed is true

TEST COMPLETE
